% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{check_activity}
\alias{check_activity}
\title{Variable activity checker}
\usage{
check_activity(data, proportion = T)
}
\arguments{
\item{data}{matrix}
}
\description{
Checks the proportion or number of active variables for each column in the dataset
(i.e. non-zero entries)
}
