% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Function to combine data}
\usage{
combine_data(list_of_data, dim)
}
\arguments{
\item{list_of_data}{the data that you wish to combine. Each element in this
list must be another list with responses y, design matrix X,
unique row counts of full data and unique row counts of
design matrix}

\item{dim}{dimension}
}
\value{
combined data
}
\description{
Function to combine data
}
