% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_GLMR_sampler.R
\name{hmc_base_sampler_GLMR}
\alias{hmc_base_sampler_GLMR}
\title{HMC sampler for base level for Poisson / NB regression model}
\usage{
hmc_base_sampler_GLMR(
  likelihood = "Poisson",
  nsamples,
  warmup,
  data_split,
  C,
  phi = NULL,
  prior_means,
  prior_variances,
  seed = sample.int(.Machine$integer.max, 1),
  n_cores = parallel::detectCores(),
  output = F
)
}
\arguments{
\item{likelihood}{the distribution for the residuals of the regression model.
Must be either "Poisson" or "NB" (Poisson by default)}

\item{nsamples}{number of samples per node}

\item{warmup}{number of burn in iterations}

\item{data_split}{list of length C where each item is a list where for c=1,...,C,
data_split[\link{c}]$y is a vector of responses and data_split[\link{c}]$X
is the design matrix}

\item{C}{number of sub-posteriors (default to 1)]}

\item{phi}{rate parameter in NB distribution (default is NULL)}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{seed}{seed number for random number generation}

\item{n_cores}{number of cores to use}

\item{output}{boolean value: defaults to T, determines whether or not
to print output to console}
}
\value{
samples from the sub-posterior targets for the split data sets
for the Poisson / NB regression model
}
\description{
Sample for base level for Poisson / NB regression
}
