% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc_GLMR_sampler.R
\name{hmc_sample_GLMR}
\alias{hmc_sample_GLMR}
\title{HMC sampler for Poisson / NB Regression model}
\usage{
hmc_sample_GLMR(
  likelihood = "Poisson",
  y,
  X,
  C,
  phi = NULL,
  prior_means,
  prior_variances,
  iterations,
  warmup,
  chains,
  seed = sample.int(.Machine$integer.max, 1),
  output = F
)
}
\arguments{
\item{likelihood}{the distribution for the residuals of the regression model.
Must be either "Poisson" or "NB" (Poisson by default)}

\item{y}{vector of responses}

\item{X}{design matrix}

\item{C}{number of sub-posterior (default to 1)}

\item{phi}{rate parameter in NB distribution (default is NULL)}

\item{prior_means}{prior for means of predictors}

\item{prior_variances}{prior for variances of predictors}

\item{iterations}{number of iterations per chain}

\item{warmup}{number of burn in iterations}

\item{chains}{number of chains}

\item{seed}{seed number for random number generation}

\item{output}{boolean value: defaults to T, determines whether or not to
print output to console}
}
\value{
samples from the (sub-)posterior target for the Poisson / NB regression model
}
\description{
Sample from (sub-)posterior using Stan
}
