% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HMCGLMR-package.R
\docType{package}
\name{HMCGLMR-package}
\alias{HMCGLMR-package}
\alias{HMCGLMR}
\title{The 'HMCGLMR' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
