% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_GLMR_data}
\alias{simulate_GLMR_data}
\title{Poisson regression data simulator}
\usage{
simulate_GLMR_data(
  N,
  likelihood,
  phi_rate = 1,
  alpha,
  frequencies,
  coefficients,
  G_means = rep(0, length(frequencies)),
  G_sds = rep(1, length(frequencies)),
  seed = NULL
)
}
\arguments{
\item{N}{number of samples}

\item{likelihood}{either "Poisson" or "NB"}

\item{phi_rate}{phi parameter for NB distribution (default is 1)}

\item{alpha}{intercept coefficient}

\item{frequencies}{frequency of coefficients occurring}

\item{coefficients}{coefficient terms}

\item{G_means}{mean of the Gaussian for simulating the data}

\item{G_sds}{standard deviations of the Gaussian for simulating the data}

\item{seed}{seed number for random number generation}
}
\value{
simulated data for Poisson regression
}
\description{
Simulate Poisson regression data
}
