% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Function to split a dataframe into C sub-dataframes}
\usage{
split_data(dataframe, y_col_index, X_col_index, C, as_dataframe = TRUE)
}
\arguments{
\item{dataframe}{data}

\item{y_col_index}{index for responses y in data frame}

\item{C}{number of subsets}

\item{as_dataframe}{determines whether or not to return the split data as dataframes
or as responses and inputs}

\item{x_col_index}{index / indices for input variables X in data frame}
}
\value{
split data
}
\description{
Function to split a dataframe into C sub-dataframes
}
