% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{unique_row_count}
\alias{unique_row_count}
\title{Function to obtain unique rows in the data}
\usage{
unique_row_count(y, X)
}
\arguments{
\item{y}{vector of responses}

\item{X}{design matrix}
}
\value{
list containing the unique rows of the full data and the unique rows
of the design matrix with their corresponding counts
}
\description{
Function to obtain unique rows in the data
}
